% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gastro_families}
\alias{gastro_families}
\title{Family labels for the built-in signatures}
\format{
Named character vector.
}
\usage{
gastro_families
}
\description{
Family labels for the built-in signatures
}
\keyword{datasets}
