% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gastro_signatures}
\alias{gastro_signatures}
\title{Published immune and stromal signatures}
\format{
Named list of character vectors.
}
\usage{
gastro_signatures
}
\description{
Published immune and stromal signatures
}
\keyword{datasets}
