% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gc_tcga}
\alias{gc_tcga}
\title{Gastric cancer prognostic cohort (TCGA-STAD)}
\format{
A list with `expr`, `time`, `event`, `stage`, `age`.
}
\source{
TCGA via UCSC Xena; endpoints from TCGA-CDR
}
\usage{
gc_tcga
}
\description{
Primary tumours with overall survival. Expression is log2 RSEM restricted to the
6,000 most variable genes.
}
\keyword{datasets}
